# hpack encoder optimizations.