… src/core/ext/transport/chttp2/transport/hpack_encoder.cc